requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.25.0
plotly>=5.15.0
pandas>=2.0.0
groq>=0.4.1
